Generating CSV for openshift
- fix operator SDK CSV markers generation
- add files from pull request 149
- generate bundle and manifest using operator SDK kustomize and OPM
- generating image bundle for test and instalation on operator hub
- add k8s resources to install in the openshift cluster
- fix required image for Openshift
