+++ PROJECT
@@ -1,0 +1,11 @@
+domain: cockroachlabs.com
+layout: go.kubebuilder.io/v2
+projectName: cockroach-operator
+repo: github.com/cockroach-operator
+resources:
+- group: crdb.cockroachlabs.com
+  kind: CrdbCluster
+  version: v1alpha1
+version: 3-alpha
+plugins:
+  go.sdk.operatorframework.io/v2-alpha: {}
