+++ api/v1alpha1/BUILD.bazel
@@ -7,9 +7,7 @@ go_library(
         "cluster_types.go",
         "condition_types.go",
         "groupversion_info.go",
-        "pod_container_types.go",
         "volume.go",
-        "volume_types.go",
         "zz_generated.deepcopy.go",
     ],
     importpath = "github.com/cockroachdb/cockroach-operator/api/v1alpha1",
