+++ api/v1alpha1/condition_types.go
@@ -16,9 +16,9 @@ limitations under the License.
 
 package v1alpha1
 
-import (
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-)
+// import (
+// 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+// )
 
 type ClusterConditionType string
 
@@ -29,16 +29,3 @@ const (
 // TODO(vladdy): this should eventually be converged with
 //     https://github.com/kubernetes/enhancements/pull/1624
 
-// ClusterCondition represents cluster status as it is perceived by
-// the operator
-type ClusterCondition struct {
-	// +required
-	// Type/Name of the condition
-	Type ClusterConditionType `json:"type"`
-	// +required
-	// Condition status: True, False or Unknown
-	Status metav1.ConditionStatus `json:"status"`
-	// +required
-	// The time when the condition was updated
-	LastTransitionTime metav1.Time `json:"lastTransitionTime"`
-}
