+++ bundle/manifests/cockroach-operator-role-default-binding_rbac.authorization.k8s.io_v1beta1_clusterrolebinding.yaml
@@ -1,0 +1,15 @@
+apiVersion: rbac.authorization.k8s.io/v1beta1
+kind: ClusterRoleBinding
+metadata:
+  creationTimestamp: null
+  labels:
+    app: cockroach-operator
+  name: cockroach-operator-role-default-binding
+roleRef:
+  apiGroup: rbac.authorization.k8s.io
+  kind: ClusterRole
+  name: cockroach-operator-role-default
+subjects:
+- kind: ServiceAccount
+  name: cockroach-operator-sa-default
+  namespace: default
