+++ bundle/manifests/cockroach-operator-sa-default_v1_serviceaccount.yaml
@@ -1,0 +1,7 @@
+apiVersion: v1
+kind: ServiceAccount
+metadata:
+  creationTimestamp: null
+  labels:
+    app: cockroach-operator
+  name: cockroach-operator-sa-default
