+++ config/crd/bases/crdb.cockroachlabs.com_crdbclusters.yaml
@@ -50,8 +50,8 @@ spec:
         metadata:
           type: object
         spec:
-          description: CrdblusterSpec defines the desired state of a CockroachDB Cluster
-            that the operator maintains.
+          description: CrdbClusterSpec defines the desired state of a CockroachDB
+            Cluster that the operator maintains.
           properties:
             additionalArgs:
               description: '(Optional) Additional command line arguments for the `cockroach`
@@ -270,11 +270,12 @@ spec:
               description: (Required) Container image information
               properties:
                 name:
-                  description: (Required) Container image with supported CockroachDB
-                    version
+                  description: '(Optional) Container image with supported CockroachDB
+                    version. This defaults to the version pinned to the operator and
+                    requires a full container and tag/sha name. For instance: cockroachdb/cockroachdb:v20.1'
                   type: string
                 pullPolicy:
-                  description: PullPolicy for the image
+                  description: PullPolicy for the image, which defaults to IfNotPresent.
                   type: string
                 pullSecret:
                   description: Secret name containing the dockerconfig to use for
@@ -288,10 +289,9 @@ spec:
                 Default: "25%"'
               type: string
             maxUnavailable:
-              description: TODO document on how you need to set the max or the min
-                The maximum number of pods that can be unavailable during a rolling
-                update. This number is set in the PodDistruptionBudget and defaults
-                to 1.
+              description: The maximum number of pods that can be unavailable during
+                a rolling update. This number is set in the PodDistruptionBudget and
+                defaults to 1.
               format: int32
               type: integer
             minUnavailable:
@@ -306,8 +306,7 @@ spec:
                 of files is expected (tls.key, tls.crt, ca.crt) Default: ""'
               type: string
             nodes:
-              description: 'Important: Run "make" to regenerate code after modifying
-                this file (Required) Number of nodes (pods) in the cluster'
+              description: (Required) Number of nodes (pods) in the cluster
               format: int32
               type: integer
             resources:
@@ -346,7 +345,6 @@ spec:
                 usage Default: false'
               type: boolean
           required:
-          - image
           - nodes
           type: object
         status:
