+++ config/crd/kustomizeconfig.yaml
@@ -1,0 +1,18 @@
+# This file is for teaching kustomize how to substitute name and namespace reference in CRD
+# This file is for teaching kustomize how to substitute name and namespace reference in CRD
+nameReference:
+  - kind: Service
+    version: v1
+    fieldSpecs:
+      - kind: CustomResourceDefinition
+        group: apiextensions.k8s.io
+        path: spec/conversion/webhookClientConfig/service/name
+
+namespace:
+  - kind: CustomResourceDefinition
+    group: apiextensions.k8s.io
+    path: spec/conversion/webhookClientConfig/service/namespace
+    create: false
+
+varReference:
+  - path: metadata/annotations
