+++ config/manifests/kustomization.yaml
@@ -1,0 +1,5 @@
+resources:
+  - ../crd
+  - ../../manifests
+  - ../samples
+  - ../scorecard
