+++ config/samples/kustomization.yaml
@@ -1,0 +1,4 @@
+## Append samples you want in your CSV to this file as resources ##
+resources:
+  - crdb-tls-example.yaml
+# +kubebuilder:scaffold:manifestskustomizesamples
