+++ hack/bin/deps.bzl
@@ -27,6 +27,8 @@ def install():
     install_kind()
     install_kubetest2()
     install_kubetest2_kind()
+    install_operator_sdk()
+    # install_opm()
 
     # Install golang.org/x/build as kubernetes/repo-infra requires it for the
     # build-tar bazel target.
@@ -271,5 +273,40 @@ def install_kubetest2_kind():
         sha256 = "e1b7ce0eec0c3db97b4fce3659e25f6190188c9c53f81ae3d090da47265a7599",
         urls = ["https://storage.googleapis.com/crdb-bazel-artifacts/linux/kubetest2-kind"],
     )
+
+## Fetch operator-sdk used on generating csv
+def install_operator_sdk():
+    # install kubetest2-kind binary
+    # TODO osx support
+    http_file(
+       name = "operator_sdk_darwin",
+       executable = 1,
+       sha256 = "7e293cd35b99c1949cbb116275bc50d7f3aa0b520fe7e53e57b09e8096e63d4e",
+       urls = ["https://github.com/operator-framework/operator-sdk/releases/download/v1.1.0/operator-sdk-v1.1.0-x86_64-apple-darwin"],
+    )
+
+    http_file(
+        name = "operator_sdk_linux",
+        executable = 1,
+        sha256 = "e0cfd1408ea8849fb32345d7f9954a2751fef7fcf4505f93db8f675d12f137ad ",
+        urls = ["https://github.com/operator-framework/operator-sdk/releases/download/v1.1.0/operator-sdk-v1.1.0-x86_64-linux-gnu"],
+    )
+#     ## Fetch opm used on generating csv
+# def install_opm():
+#     http_file(
+#        name = "opm_darwin",
+#        executable = 1,
+#        sha256 = "7e293cd35b99c1949cbb116275bc50d7f3aa0b520fe7e53e57b09e8096e63d4e",
+#        urls = ["https://storage.googleapis.com/crdb-bazel-artifacts/macos/opm"],
+#     )
+
+#     http_file(
+#         name = "opm_linux",
+#         executable = 1,
+#         sha256 = "e0cfd1408ea8849fb32345d7f9954a2751fef7fcf4505f93db8f675d12f137ad ",
+#         urls = ["https://storage.googleapis.com/crdb-bazel-artifacts/linux/opm"],
+#     )
+
+
