+++ index.Dockerfile389314389
@@ -1,0 +1,6 @@
+FROM quay.io/operator-framework/upstream-opm-builder
+LABEL operators.operatorframework.io.index.database.v1=/database/index.db
+ADD index_build_tmp162048078/database/index.db /database/index.db
+EXPOSE 50051
+ENTRYPOINT ["/bin/opm"]
+CMD ["registry", "serve", "--database", "/database/index.db"]
