+++ openshift_automation_drafts.txt
@@ -1,0 +1,28 @@
+ ./hack/update-crds.sh --> generateCRD
+operator-sdk generate kustomize manifests --verbose --> GENERATE CSV base
+cp config/manifests/bases/cockroach-operator.clusterserviceversion.yaml  deploy/bases/
+cp config/crd/bases/crdb.cockroachlabs.com_crdbclusters.yaml deploy/crds/crdb.cockroachlabs.com_crdbclusters.yaml
+cp config/rbac/role.yaml deploy/role.yaml
+
+operator-sdk generate packagemanifests -q --version 0.0.1 --input-dir  deploy  --crds-dir  deploy --deploy-dir manifests
+
+
+
+ ./hack/update-crds.sh --> generateCRD
+operator-sdk generate kustomize manifests -q
+cd config/manager && $(KUSTOMIZE) edit set image controller=$(IMG)
+kustomize build config/manifests | operator-sdk generate bundle -q --overwrite --version $(VERSION) $(BUNDLE_METADATA_OPTS)
+operator-sdk bundle validate ./bundle
+
+
+docker build -t quay.io/alinalion/cockroach-operator:latest -f bundle.Dockerfile .
+docker build -t quay.io/alinalion/cockroach-operator:0.0.3 -f bundle.Dockerfile .
+docker push quay.io/alinalion/cockroach-operator:latest 
+docker push quay.io/alinalion/cockroach-operator:0.0.3 
+
+./opm index add --bundles quay.io/alinalion/cockroach-operator:0.0.3 --tag quay.io/alinalion/cockroach-operator:index-latest -c docker
+docker push quay.io/alinalion/cockroach-operator:index-latest
+
+./opm alpha bundle generate --directory ./bundle/manifests --package cockroach-operator \
+--channels alpha --default alpha
+
